Split oversized accord_home part files into smaller parts
accord_home_message_row.dart (1200 lines) and accord_home_rail.dart
(1124 lines) were the only first-party source files over 1000 lines.
Break them up along class boundaries into new part files of the same
accord_home.dart library, with no code changes:

- accord_home_mute_button.dart: _MuteButton + _NotifAction (channel
  header notification toggle, unrelated to message rows)
- accord_home_reactions.dart: _ReactButton, _ReactionPill,
  _ReactorsDialog
- accord_home_space_actions.dart: _serverActionEntries,
  _copyServerLink, _leaveSpace, _leaveAndDeleteSpace
- accord_home_rail_tiles.dart: _DraggableSpace, _FolderTile,
  _SpaceIcon, _DirectMessagesButton, _AddServerButton,
  _HiddenServersButton
